Work for 26.11.2023

+++ app/Livewire/Application/Sheet/List/ListSheet.php
@@ -4,6 +4,10 @@ namespace App\Livewire\Application\Sheet\List;
 
 use App\Models\ConsultationSheet;
 use App\Models\Hospital;
+use App\Repositories\Sheet\Get\GetConsultationSheetRepository;
+use Illuminate\Contracts\Foundation\Application;
+use Illuminate\Contracts\View\Factory;
+use Illuminate\Contracts\View\View;
 use Livewire\Attributes\Url;
 use Livewire\Component;
 use Livewire\WithPagination;
@@ -25,7 +29,11 @@ class ListSheet extends Component
     #[Url(as: 'sortAsc')]
     public $sortAsc = true;
 
-
+    /**
+     * Get selected Subscription when selectedIndex listener emitted in parent view
+     * @param int $selectedIndex
+     * @return void
+     */
     public function getSelectedIndex(int $selectedIndex): void
     {
         $this->selectedIndex=$selectedIndex;
@@ -37,31 +45,59 @@ class ListSheet extends Component
         $this->dispatch('sheetInfo');
         $this->dispatch('open-form-new');
     }
+    /**
+     * Open form sheet modal in creation mode
+     * @param ConsultationSheet $consultationSheet
+     * @return void
+     */
     public  function  newRequestForm(ConsultationSheet $consultationSheet): void
     {
         $this->dispatch('open-new-request-form');
         $this->dispatch('consultationSheet',$consultationSheet);
     }
+    /**
+     * Open form Sheet modal in editable mode
+     * @param ConsultationSheet $sheet
+     * @return void
+     */
     public  function  edit(ConsultationSheet $sheet): void
     {
         $this->dispatch('open-form-new');
         $this->dispatch('sheetInfo',$sheet);
         $this->dispatch('selectedIndex',$this->selectedIndex);
     }
+    /**
+     * Show delete sheet dialog
+     * @param ConsultationSheet $sheet
+     * @return void
+     */
     public function showDeleteDialog(ConsultationSheet $sheet): void
     {
         $this->sheet=$sheet;
         $this->dispatch('delete-sheet-dialog');
     }
+    /**
+     * Destroy Consultation sheet
+     * @return void
+     */
     public function delete(): void
     {
         try {
-            $this->sheet->delete();
-            $this->dispatch('sheet-deleted', ['message' => "Fiche de consultation bien rétiré !"]);
+            if ($this->sheet->consultationRequests->isEmpty()){
+                $this->sheet->delete();
+                $this->dispatch('sheet-deleted', ['message' => "Fiche de consultation bien rétiré !"]);
+            }else{
+                $this->dispatch('error', ['message' => 'Action impossible, SVP, Cette fiche contient pluesieus données']);
+            }
         }catch (\Exception $ex){
             $this->dispatch('error', ['message' => $ex->getMessage()]);
         }
     }
+    /**
+     * Sort consultation Sheet (ASC or DESC)
+     * @param $value
+     * @return void
+     */
     public function sortSheet($value): void
     {
         if($value==$this->sortBy){
@@ -69,30 +105,28 @@ class ListSheet extends Component
         }
         $this->sortBy = $value;
     }
-
+    /**
+     * Mounted component
+     * @param int $selectedIndex
+     * @return void
+     */
     public  function mount(int $selectedIndex): void
     {
         $this->selectedIndex=$selectedIndex;
-
     }
-
+    /**
+     * Render component
+     * @return Application|Factory|View|\Illuminate\Foundation\Application
+     */
     public function render()
     {
         return view('livewire.application.sheet.list.list-sheet',[
-            'sheets'=>ConsultationSheet::join('subscriptions','subscriptions.id','consultation_sheets.subscription_id')
-                ->where('consultation_sheets.subscription_id',$this->selectedIndex)
-                ->when($this->q, function ($query) {
-                    return $query->where(function ($query) {
-                        return $query->where('consultation_sheets.name', 'like', '%' . $this->q . '%')
-                            ->orWhere('consultation_sheets.email', 'like', '%' . $this->q . '%')
-                            ->orWhere('consultation_sheets.number_sheet', 'like', '%' . $this->q . '%')
-                            ->orWhere('consultation_sheets.phone', 'like', '%' . $this->q . '%')
-                            ->orWhere('consultation_sheets.registration_number', 'like', '%' . $this->q . '%');
-                    });
-                })->orderBy($this->sortBy, $this->sortAsc ? 'ASC' : 'DESC')
-                ->select('consultation_sheets.*','subscriptions.name as subscription')
-                ->where('consultation_sheets.hospital_id',Hospital::DEFAULT_HOSPITAL)
-                ->paginate(10)
+            'sheets'=>GetConsultationSheetRepository::getConsultationSheetList(
+                $this->selectedIndex,
+                $this->q,
+                $this->sortBy,
+                $this->sortAsc
+            )
         ]);
     }
 }
